Add testing for C++-based exception handling
