+++ src/tests/Common/dirs.proj
@@ -14,7 +14,7 @@
       <DisabledProjects Include="$(TestRoot)Loader\classloader\StaticVirtualMethods\**\generatetest.csproj" /> <!-- test generators -->
       <DisabledProjects Include="$(TestRoot)Performance\Scenario\JitBench\unofficial_dotnet\JitBench.csproj" /> <!-- no official build support for SDK-style netcoreapp2.0 projects -->
       <DisabledProjects Include="$(TestRoot)TestWrappers*\**\*.csproj" />
-      <DisabledProjects Include="$(TestRoot)nativeaot\SmokeTests\HelloWasm\HelloWasm.csproj" Condition="!('$(TestBuildMode)' == 'nativeaot' and '$(TargetArchitecture)' == 'wasm')"/>
+      <DisabledProjects Include="$(TestRoot)nativeaot\SmokeTests\HelloWasm\*.csproj" Condition="!('$(TestBuildMode)' == 'nativeaot' and '$(TargetArchitecture)' == 'wasm')"/>
       <DisabledProjects Include="$(TestRoot)nativeaot\**\*.csproj" Condition="'$(TestBuildMode)' != 'nativeaot'" />
     </ItemGroup>
 
@@ -36,6 +36,7 @@
       <DisabledProjects Include="$(TestRoot)nativeaot\SmokeTests\TrimmingBehaviors\*.csproj" Condition="'$(TargetOS)' == 'wasi'" />
       <DisabledProjects Include="$(TestRoot)nativeaot\SmokeTests\UnitTests\*.csproj" Condition="'$(TargetOS)' == 'wasi'" />
 
+      <DisabledProjects Include="$(TestRoot)nativeaot\SmokeTests\HelloWasm\ExceptionHandlingTests.Cpp.csproj" Condition="'$(TargetOS)' == 'wasi'" />
       <DisabledProjects Include="$(TestRoot)nativeaot\SmokeTests\HelloWasm\WasmDebugging.csproj" />
     </ItemGroup>
