+++ src/tests/nativeaot/SmokeTests/HelloWasm/ExceptionHandlingTests.Cpp.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <CLRTestKind>BuildAndRun</CLRTestKind>
+    <CLRTestPriority>0</CLRTestPriority>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+
+    <!-- Test that the C++-based exception dispatch is working. -->
+    <WasmEnableExceptionHandling>false</WasmEnableExceptionHandling>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="ExceptionHandlingTests.Cpp.cs" />
+    <Compile Include="ExceptionHandlingTests.Common.cs" />
+  </ItemGroup>
+</Project>
